CI: Add release workflow
This github actions workflow will create a draft release for PRs that
updates the version file.

The commit message subject line must be
Release microcom <version>

The commit message body will be used as release description. The release
is converted from draft to an actual release, and the commit thereby
tagged once the PR is merged.

Signed-off-by: Jonas Rebmann <jre@pengutronix.de>